pyproject.toml
[project]
name = "cdr-booklet-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-sqlalchemy>=3.1.1",
    "python-dotenv>=1.2.1",
    "flask-jwt-extended>=4.7.1",
    "gunicorn>=23.0.0",
    "celery>=5.5.3",
    "flower>=2.0.1",
    "redis>=7.1.0",
    "pytest>=9.0.1",
]

[project.dependency-groups]
dev = [
    "ruff>=0.8.0",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pycache__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "logs",
    "venv",
]

# Same as Black.
line-length = 100
indent-width = 4

# Assume Python 3.12+
target-version = "py312"

[tool.ruff.lint]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[dependency-groups]
dev = [
    "ruff>=0.14.5",
]

[project.scripts]
lint = "ruff check ."
lint-fix = "ruff check . --fix"
format = "ruff format ."
sort = "ruff check . --select I --fix"
type-check = "mypy ."
